Split out code into new files
