+++ src/main.rs
@@ -9,6 +9,8 @@ use tokio::sync::mpsc::{unbounded_channel, UnboundedReceiver};
 use warp::Filter;
 
 mod config;
+mod handlers;
+mod routes;
 
 #[derive(Debug, Clone, Deserialize)]
 enum Message {
@@ -128,72 +130,6 @@ impl Controller {
             .await?;
 
         Ok(())
-    }
-}
-
-mod routes {
-    use crate::handlers;
-    use crate::Message;
-    use tokio::sync::mpsc::UnboundedSender;
-    use warp::Filter;
-
-    pub(crate) fn build(
-        tx: UnboundedSender<Message>,
-    ) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
-        trigger(tx.clone()).or(webhook(tx.clone()))
-    }
-
-    /// POST /api/trigger
-    pub(crate) fn trigger(
-        tx: UnboundedSender<Message>,
-    ) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
-        warp::path!("trigger")
-            .and(warp::post())
-            .and(with_inbox(tx))
-            .and_then(handlers::handle_trigger)
-    }
-
-    /// POST /api/webhook
-    pub(crate) fn webhook(
-        tx: UnboundedSender<Message>,
-    ) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
-        warp::path!("webhook")
-            .and(warp::post())
-            // .and(json_body())
-            .and(with_inbox(tx))
-            .and_then(handlers::handle_webhook)
-    }
-
-    fn with_inbox(
-        tx: UnboundedSender<Message>,
-    ) -> impl Filter<Extract = (UnboundedSender<Message>,), Error = std::convert::Infallible> + Clone
-    {
-        warp::any().map(move || tx.clone())
-    }
-
-    fn json_body() -> impl Filter<Extract = (Message,), Error = warp::Rejection> + Clone {
-        warp::body::json()
-    }
-}
-
-mod handlers {
-    use crate::Message;
-    use std::convert::Infallible;
-    use tokio::sync::mpsc::UnboundedSender;
-    use warp::http::StatusCode;
-
-    pub(crate) async fn handle_trigger(
-        tx: UnboundedSender<Message>,
-    ) -> Result<impl warp::Reply, Infallible> {
-        tx.send(Message::Trigger).unwrap();
-
-        Ok(StatusCode::NO_CONTENT)
-    }
-
-    pub(crate) async fn handle_webhook(
-        _tx: UnboundedSender<Message>,
-    ) -> Result<impl warp::Reply, Infallible> {
-        Ok(StatusCode::NO_CONTENT)
     }
 }
